moving qr code utils to it's own file
