+++ src/utils.ts
@@ -1,90 +1,5 @@
-// Note: have to patch _qrcode
-import _qrcode from 'qrcode-generator';
-
 import Reactotron from 'reactotron-react-native';
-import {u8aConcat} from '@polkadot/util';
-import {FRAME_SIZE} from './constants';
 import {FunctionMetadataLatest} from '@polkadot/types/interfaces';
-
-const MULTIPART = new Uint8Array([0]);
-
-// A small hurdle to jump through, just to get the default/default correct (as generated)
-export const QrCode: typeof _qrcode = _qrcode;
-
-// HACK The default function take string -> number[], the Uint8array is compatible
-// with that signature and the use thereof
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-(QrCode as any).stringToBytes = (data: Uint8Array): Uint8Array => data;
-
-export function encodeNumber(value: number): Uint8Array {
-  // eslint-disable-next-line no-bitwise
-  return new Uint8Array([value >> 8, value & 0xff]);
-}
-
-export function createFrames(input: Uint8Array): Uint8Array[] {
-  const frames = [];
-  let idx = 0;
-
-  while (idx < input.length) {
-    frames.push(input.subarray(idx, idx + FRAME_SIZE));
-
-    idx += FRAME_SIZE;
-  }
-
-  return frames.map(
-    (frame, index: number): Uint8Array => u8aConcat(MULTIPART, encodeNumber(frames.length), encodeNumber(index), frame),
-  );
-}
-
-/*
- * @return strippedData: the rawBytes from react-native-camera, stripped of the ec11 padding to fill the frame size. See: decoders.js
- * N.B. Substrate oversized/multipart payloads will already be hashed at this point.
- */
-export function rawDataToU8A(rawData: string): Uint8Array | null {
-  if (!rawData) {
-    return null;
-  }
-
-  // Strip filler bytes padding at the end
-  if (rawData.substr(-2) === 'ec') {
-    rawData = rawData.substr(0, rawData.length - 2);
-  }
-
-  while (rawData.substr(-4) === 'ec11') {
-    rawData = rawData.substr(0, rawData.length - 4);
-  }
-
-  // Verify that the QR encoding is binary and it's ending with a proper terminator
-  if (rawData.substr(0, 1) !== '4' || rawData.substr(-1) !== '0') {
-    return null;
-  }
-
-  // Strip the encoding indicator and terminator for ease of reading
-  rawData = rawData.substr(1, rawData.length - 2);
-
-  const length8 = parseInt(rawData.substr(0, 2), 16) || 0;
-  const length16 = parseInt(rawData.substr(0, 4), 16) || 0;
-  let length = 0;
-
-  // Strip length prefix
-  if (length8 * 2 + 2 === rawData.length) {
-    rawData = rawData.substr(2);
-    length = length8;
-  } else if (length16 * 2 + 4 === rawData.length) {
-    rawData = rawData.substr(4);
-    length = length16;
-  } else {
-    return null;
-  }
-
-  const bytes = new Uint8Array(length);
-
-  for (let i = 0; i < length; i++) {
-    bytes[i] = parseInt(rawData.substr(i * 2, 2), 16);
-  }
-
-  return bytes;
-}
 
 export const ReactotronDebug = Reactotron.debug;
